+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 dd mm yyyy
 
+- Fixed and improved `/universe/names/` ESI requests.
 - Some refactoring of the frontend, which also slightly changes the behavior when changing system settings.
 - Small text improvements/fixes.